% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CensoEducacaoSuperior_dics}
\alias{CensoEducacaoSuperior_dics}
\title{Censo da Educação Superior dictionaries}
\format{A list with 18 lists (one for each year), each one of these lists contains 11 data frames (one for each file type), each data frame is named 'dic_filetype_year'. The general data frame contains 9 variables and one observation for each variable of the dataset that is used to import:
\describe{
  \item{int_pos}{start position of the variable in each line}
  \item{var_name}{name of the variable}
  \item{label}{descriptive label for the variable, if available}
  \item{length}{length of the variable}
  \item{decimal_places}{number of decimal places of the variable}
  \item{fin_pos}{final position of the variable in each line}
  \item{col_types}{suggested column type of the variable. 'c' for character, 'i' for integer and 'd' for double}
  \item{CHAR}{dummy variable for character type, used to ease programming with}
  ...
}}
\source{
\url{http://portal.inep.gov.br/basica-levantamentos-acessar}
}
\usage{
CensoEducacaoSuperior_dics
}
\description{
A list containing INEP import dictionaries for Censo da Educação Superior from 1995 to 2012 translated into data frames.
}
\examples{
\dontrun{ data(CensoEducacaoSuperior_dics)

## Acess ies dictionary for 2010:
CensoEducacaoSuperior_dics$`2010`$`dic_ies_2010`

## Alternatively, use the wrapper function
get_import_dictionary("CensoEducacaoSuperior", 2010, "ies")}

}
\keyword{datasets}

