% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.r
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Returns available datasets int the package}
\usage{
get_available_datasets()
}
\description{
Returns available datasets int the package
}
