% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.r
\name{get_available_filetypes}
\alias{get_available_filetypes}
\title{Returns available filetypes for a dataset in a given period.}
\usage{
get_available_filetypes(dataset, period)
}
\arguments{
\item{dataset}{name of the dataset. See get_available_datasets() for options.}

\item{period}{See get_available_periods(dataset) for options}
}
\description{
Returns available filetypes for a dataset in a given period.
}
\examples{

get_available_filetypes("PNAD", 2014)

}
