% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.r
\name{get_available_periods}
\alias{get_available_periods}
\title{Returns the periods for wich we have information about a dataset in the package}
\usage{
get_available_periods(dataset, fwfonly = FALSE)
}
\arguments{
\item{dataset}{name of the dataset. See get_available_datasets() for options.}

\item{fwfonly}{(optional) TRUE/FALSE if TRUE returns only the periods for wich the dataset is distributed as a fixed width file.}
}
\description{
Returns the periods for wich we have information about a dataset in the package
}
