% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_functions.r
\name{get_import_dictionary}
\alias{get_import_dictionary}
\title{Returns fwf dictionary}
\usage{
get_import_dictionary(dataset, i, ft)
}
\arguments{
\item{dataset}{name of the dataset, between quotes. If a wrong name is set in this parameter you will see a list of the available datasets.}

\item{i}{period, usually year}

\item{ft}{file type.}
}
\description{
The dictionaries used in this package are stored in lists. get_import_dictionary() is a wrapper to search dictionaries based on dataset name, period and file type.
}
\examples{
get_import_dictionary(dataset = "CensoEscolar", i = 1996, ft = "escola")
}
