% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.r
\name{nodic_overlap}
\alias{nodic_overlap}
\title{Function to search and solve overlaping problems, returning a list of dictionaries.}
\usage{
nodic_overlap(dic, init_pos = "int_pos", fin_pos = "fin_pos")
}
\arguments{
\item{dic}{The dictionary to be worked on.}

\item{init_pos}{The name of the column which contains the initial positions.}

\item{fin_pos}{The name of the column which contains the final positions.}
}
\description{
Function to search and solve overlaping problems, returning a list of dictionaries.
}
