% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.r
\name{parse_SAS_import_dic}
\alias{parse_SAS_import_dic}
\title{transforms IBGE's SAS inputs into dataframe dic}
\usage{
parse_SAS_import_dic(file, keepLabels = FALSE)
}
\arguments{
\item{file}{SAS input file}

\item{keepLabels}{keep variables labels?}
}
\description{
transforms IBGE's SAS inputs into dataframe dic
}
