% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PME_dics}
\alias{PME_dics}
\title{Pesquisa Mensal de Emprego dictionaries}
\format{A list with 1 list (one for each year), each one of these lists contains 2 data frames (one for each file type), each data frame is named 'dic_filetype_year'. The general data frame contains 9 variables and one observation for each variable of the dataset that is used to import:
\describe{
  \item{int_pos}{start position of the variable in each line}
  \item{var_name}{name of the variable}
  \item{label}{descriptive label for the variable, if available}
  \item{length}{length of the variable}
  \item{decimal_places}{number of decimal places of the variable}
  \item{fin_pos}{final position of the variable in each line}
  \item{col_types}{suggested column type of the variable. 'c' for character, 'i' for integer and 'd' for double}
  \item{CHAR}{dummy variable for character type, used to ease programming with}
  ...
}}
\source{
\url{http://www.ibge.gov.br/home/estatistica/indicadores/trabalhoerendimento/pme_nova/defaultmicro.shtm}
}
\usage{
data(PME_dics)

## Acess escolas dictionary for 2000:
(PME_dics$`2012.1`$`dic_PME_2012.1`)

## Alternatively, use the wrapper function
get_import_dictionary("PME", 2012.1, "PME")
}
\description{
A list containing IBGE import dictionaries for PME from 2002.1 to 2015.2 translated into data frames.
}
\keyword{datasets}

