% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PNAD_dics}
\alias{PNAD_dics}
\title{PNAD dictionaries}
\format{A list with 13 lists (one for each year), each one of these lists contains 2 data frames (one for each file type), each data frame is named 'dic_filetype_year'. The general data frame contains 9 variables and one observation for each variable of the dataset that is used to import:
\describe{
  \item{int_pos}{start position of the variable in each line}
  \item{var_name}{name of the variable}
  \item{label}{descriptive label for the variable, if available}
  \item{length}{length of the variable}
  \item{decimal_places}{number of decimal places of the variable}
  \item{fin_pos}{final position of the variable in each line}
  \item{col_types}{suggested column type of the variable. 'c' for character, 'i' for integer and 'd' for double}
  \item{CHAR}{dummy variable for character type, used to ease programming with}
  ...
}}
\source{
\url{http://www.ibge.gov.br/home/estatistica/populacao/trabalhoerendimento/pnad2014/microdados.shtm}
}
\usage{
data(PNAD_dics)

## Acess pessoas dictionary for 2014:
PNAD_dics$`2014`$`dic_pessoas_2014`

## Alternatively, use the wrapper function
get_import_dictionary("PNAD", 2014, "pessoas")
}
\description{
A list containing IBGE dictionaries for PNAD from 2001 to 2014 translated into data frames.
}
\keyword{datasets}

