% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PNADcontinua_dics}
\alias{PNADcontinua_dics}
\title{PNAD continua dictionaries}
\format{A list with 17 lists (one for each year), each one of these lists contains 1 data frame (one for each file type), each data frame is named 'dic_filetype_year'. The general data frame contains 9 variables and one observation for each variable of the dataset that is used to import:
\describe{
  \item{int_pos}{start position of the variable in each line}
  \item{var_name}{name of the variable}
  \item{label}{descriptive label for the variable, if available}
  \item{length}{length of the variable}
  \item{decimal_places}{number of decimal places of the variable}
  \item{fin_pos}{final position of the variable in each line}
  \item{col_types}{suggested column type of the variable. 'c' for character, 'i' for integer and 'd' for double}
  \item{CHAR}{dummy variable for character type, used to ease programming with}
  ...
}}
\source{
\url{http://www.ibge.gov.br/home/estatistica/indicadores/trabalhoerendimento/pnad_continua/default_microdados.shtm}
}
\usage{
PNADcontinua_dics
}
\description{
A list containing IBGE import dictionaries for PNAD Continua from 2012.1 to 2016.1 translated into data frames.
}
\examples{
\dontrun{data(PNADcontinua_dics)

## Acess pessoas dictionary for 2012.1:
(PNADcontinua_dics$`2012.1`$`dic_pessoas_2012.1`)

## Alternatively, use the wrapper function
get_import_dictionary("PNADcontinua", "2012.1", "pessoas")}

}
\keyword{datasets}

