% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POF_dics}
\alias{POF_dics}
\title{Pesquisa de Orçamentos Familiares dictionaries}
\format{A list with 1 list (one for each year), each one of these lists contains 16 data frame (one for each file type), each data frame is named 'dic_filetype_year'. The general data frame contains 9 variables and one observation for each variable of the dataset that is used to import:
\describe{
  \item{int_pos}{start position of the variable in each line}
  \item{var_name}{name of the variable}
  \item{label}{descriptive label for the variable, if available}
  \item{length}{length of the variable}
  \item{decimal_places}{number of decimal places of the variable}
  \item{fin_pos}{final position of the variable in each line}
  \item{col_types}{suggested column type of the variable. 'c' for character, 'i' for integer and 'd' for double}
  \item{CHAR}{dummy variable for character type, used to ease programming with}
  ...
}}
\source{
\url{http://www.ibge.gov.br/home/estatistica/populacao/condicaodevida/pof/2008_2009/microdados.shtm}
}
\usage{
POF_dics
}
\description{
A list containing IBGE import dictionaries for POF from 2008/2009 translated into data frames.
}
\examples{
 \dontrun{data(POF_dics)

## Acess T_DESPESA_90DIAS_S dictionary for 2008:
POF_dics$`2008`$`dic_T_DESPESA_90DIAS_S_2008`

## Alternatively, use the wrapper function
get_import_dictionary("POF", "2008", "T_DESPESA_90DIAS_S")}

}
\keyword{datasets}

