% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_import_functions.R
\name{read_data_censoescolar_ff}
\alias{read_data_censoescolar_ff}
\title{Reads Censo Escolar .csv files with ff}
\usage{
read_data_censoescolar_ff(ft, i, vars_subset = NULL, root_path)
}
\arguments{
\item{ft}{file type. Indicates the subdataset within the dataset. For example: "pessoa" (person) or "domicílio" (household) data from the "CENSO" (Census). For a list of available ft for the period just type an invalid ft (Ex: ft = 'aasfasf')}

\item{i}{period: yyyy format}

\item{vars_subset}{string vector -- each element refers to a dataframe column}

\item{root_path}{a path to the directory where dataset was downloaded}
}
\description{
Reads Censo Escolar .csv files with ff
}
