% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wrapper_functions.R
\name{read_dataset}
\alias{read_CAGED}
\alias{read_CENSO}
\alias{read_CensoEducacaoSuperior}
\alias{read_CensoEscolar}
\alias{read_PME}
\alias{read_PNAD}
\alias{read_PNADcontinua}
\alias{read_POF}
\alias{read_RAIS}
\alias{read_SISOB}
\alias{read_dataset}
\title{Functions to read most common Brazillian dataset}
\usage{
read_CensoEscolar(ft, i, harmonize_varnames = F, root_path = NULL,
  file = file)

read_CensoEducacaoSuperior(ft, i, root_path = NULL, file = NULL)

read_SISOB(ft, i, root_path = NULL, file = NULL)

read_CENSO(ft, i, root_path = NULL, file = NULL, UF = NULL)

read_RAIS(ft, i, root_path = NULL, file = NULL, UF = NULL)

read_CAGED(ft, i, root_path = NULL, file = NULL)

read_PNAD(ft, i, root_path = NULL, file = NULL)

read_PME(ft, i, root_path = NULL, file = NULL)

read_POF(ft, i, root_path = NULL, file = NULL)

read_PNADcontinua(ft, i, root_path = NULL, file = NULL)
}
\arguments{
\item{ft}{file type. Indicates the subdataset within the dataset. For example: "pessoa" (person) or "domicílio" (household) data from the "CENSO" (Census). For a list of available ft for the period just type an invalid ft (Ex: ft = 'aasfasf')}

\item{i}{period. Normally period in YYY format.}

\item{harmonize_varnames}{Should variable names be harmonized over the periods of the subdataset (ft)}
}
\value{
a data.frame containing the imported data.
}
\description{
The funtions read_DatasetName read each dataset easily and quickly. The functions gather the import parameters from the DatasetName_metadata_file_harmonization.csv files (availabe at folder extdata) and import dictionaries (available at folder data). Base on this dictionaries and parameters it dispatches the read_data function that reads the file
}

