% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_import_functions.R
\name{read_metadata}
\alias{read_metadata}
\title{Return a data.frame with metadata about a dataset}
\usage{
read_metadata(dataset)
}
\arguments{
\item{dataset}{name of the dataset. Use get_available_datasets() to see options.}
}
\value{
a data.frame containing metadata about the dataset.
}
\description{
Return a data.frame with metadata about a dataset
}

