% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{unzip_all_7z_rar}
\alias{unzip_all_7z_rar}
\title{Wrapper for unzipping lots of .rar and .7z files with archive::archive() .}
\usage{
unzip_all_7z_rar(root_path)
}
\arguments{
\item{root_path}{path to the directory where files are stored(will look for files recursively inside that folder)}
}
\description{
Wrapper for unzipping lots of .rar and .7z files with archive::archive() .
}
\examples{
\dontrun{

unzip_all_7z_rar("Datasets/micro_censo_escolar_2010")

}

}
