% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_import_functions.R
\name{aux_read_fwf}
\alias{aux_read_fwf}
\title{Reads fixed-width file (fwf) file based on dictionary.}
\usage{
aux_read_fwf(f, dic)
}
\arguments{
\item{f}{A fixed-width file (fwf) (normally a .txt file).}

\item{dic}{A data.frame, containing the import dictionary, including the variables: var_name, int_pos, fin_pos, decima_places (optional)}
}
\value{
a data.frame containing the imported data.
}
\description{
Reads fixed-width file (fwf) file based on dictionary.
}
\examples{
aux_read_fwf(filename.txt,dicionary_name)
}

