% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wrapper_functions.R
\name{read_dataset}
\alias{read_CAGED}
\alias{read_CENSO}
\alias{read_CensoEducacaoSuperior}
\alias{read_CensoEscolar}
\alias{read_ENEM}
\alias{read_PME}
\alias{read_PNAD}
\alias{read_PNADcontinua}
\alias{read_PNS}
\alias{read_POF}
\alias{read_RAIS}
\alias{read_dataset}
\title{Functions to read most common Brazillian dataset}
\usage{
read_CensoEscolar(ft, i, harmonize_varnames = F, root_path = NULL,
  file = NULL, vars_subset = NULL, nrows = -1L, source_file_mark = F)

read_CensoEducacaoSuperior(ft, i, root_path = NULL, file = NULL,
  vars_subset = NULL, nrows = -1L, source_file_mark = F)

read_ENEM(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  nrows = -1L, source_file_mark = F)

read_CENSO(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  UF = NULL, nrows = -1L, source_file_mark = F)

read_RAIS(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  UF = NULL, nrows = -1L, source_file_mark = F)

read_CAGED(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  nrows = -1L, source_file_mark = F)

read_PNAD(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  nrows = -1L, source_file_mark = F)

read_PME(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  nrows = -1L, source_file_mark = F)

read_POF(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  nrows = -1L, source_file_mark = F)

read_PNADcontinua(ft, i, root_path = NULL, file = NULL,
  vars_subset = NULL, nrows = -1L, source_file_mark = F)

read_PNS(ft, i, root_path = NULL, file = NULL, vars_subset = NULL,
  nrows = -1L, source_file_mark = F)
}
\arguments{
\item{ft}{file type. Indicates the subdataset within the dataset. For example: "pessoa" (person) or "domicílio" (household) data from the "CENSO" (Census). For a list of available ft for the period just type an invalid ft (Ex: ft = 'aasfasf')}

\item{i}{period. Normally period in YYY format.}

\item{harmonize_varnames}{Should variable names be harmonized over the periods of the subdataset (ft)}

\item{root_path}{(optional) a path to the directory where dataset was downloaded}

\item{file}{(optional) file to read, ignore all metadata in this case}

\item{vars_subset}{(optional) read only selected variables( named on the dictionary for fwf files or in the first row for delimited files)}

\item{nrows}{(optional) read only n first rows}

\item{source_file_mark}{(optional) TRUE/FALSE , if T create a variable with the filename that the observation was imported from, useful for datasets with lots of separated files( CENSO and RAIS)}

\item{UF}{(optional) only for CENSO and RAIS. Use this option to read only the files for selected brazilian states. c("DF)}
}
\value{
a data.frame containing the imported data.
}
\description{
The funtions read_DatasetName read each dataset easily and quickly. The functions gather the import parameters from the DatasetName_metadata_file_harmonization.csv files (availabe at folder extdata) and import dictionaries (available at folder data). Base on this dictionaries and parameters it dispatches the read_data function that reads the file
}

